Move the logic from LogoutHandler to the LogoutSuccessListener
