+++ core-bundle/tests/EventListener/Security/LogoutSuccessListenerTest.php
@@ -12,17 +12,37 @@ declare(strict_types=1);
 
 namespace Contao\CoreBundle\Tests\EventListener\Security;
 
+use Contao\BackendUser;
 use Contao\CoreBundle\EventListener\Security\LogoutSuccessListener;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\CoreBundle\Routing\ScopeMatcher;
-use PHPUnit\Framework\TestCase;
+use Contao\CoreBundle\Tests\TestCase;
+use Contao\System;
+use Psr\Log\LoggerInterface;
+use Scheb\TwoFactorBundle\Security\Authentication\Token\TwoFactorToken;
+use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
+use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Http\Event\LogoutEvent;
 use Symfony\Component\Security\Http\HttpUtils;
 
 class LogoutSuccessListenerTest extends TestCase
 {
+    use ExpectDeprecationTrait;
+
+    protected function tearDown(): void
+    {
+        unset($GLOBALS['TL_USERNAME']);
+
+        parent::tearDown();
+    }
+
     public function testReturnsIfResponseIsAlreadySet(): void
     {
         $response = new Response();
@@ -30,7 +50,7 @@ class LogoutSuccessListenerTest extends TestCase
         $event = new LogoutEvent(new Request(), null);
         $event->setResponse($response);
 
-        $listener = new LogoutSuccessListener($this->createMock(HttpUtils::class), $this->createMock(ScopeMatcher::class));
+        $listener = $this->mockLogoutSuccessListener();
         $listener($event);
 
         $this->assertSame($response, $event->getResponse());
@@ -58,7 +78,7 @@ class LogoutSuccessListenerTest extends TestCase
 
         $event = new LogoutEvent($request, null);
 
-        $listener = new LogoutSuccessListener($httpUtils, $scopeMatcher);
+        $listener = $this->mockLogoutSuccessListener($httpUtils, $scopeMatcher);
         $listener($event);
 
         $response = $event->getResponse();
@@ -89,7 +109,7 @@ class LogoutSuccessListenerTest extends TestCase
 
         $event = new LogoutEvent($request, null);
 
-        $listener = new LogoutSuccessListener($httpUtils, $scopeMatcher);
+        $listener = $this->mockLogoutSuccessListener($httpUtils, $scopeMatcher);
         $listener($event);
 
         $response = $event->getResponse();
@@ -120,7 +140,7 @@ class LogoutSuccessListenerTest extends TestCase
 
         $event = new LogoutEvent($request, null);
 
-        $listener = new LogoutSuccessListener($httpUtils, $scopeMatcher);
+        $listener = $this->mockLogoutSuccessListener($httpUtils, $scopeMatcher);
         $listener($event);
 
         $response = $event->getResponse();
@@ -150,12 +170,272 @@ class LogoutSuccessListenerTest extends TestCase
 
         $event = new LogoutEvent($request, null);
 
-        $listener = new LogoutSuccessListener($httpUtils, $scopeMatcher);
+        $listener = $this->mockLogoutSuccessListener($httpUtils, $scopeMatcher);
         $listener($event);
 
         $response = $event->getResponse();
 
         $this->assertInstanceOf(RedirectResponse::class, $response);
         $this->assertSame('contao_backend_login', $response->getTargetUrl());
+    }
+
+    public function testAddsTheLogEntry(): void
+    {
+        $framework = $this->mockContaoFramework();
+
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects($this->once())
+            ->method('info')
+            ->with('User "foobar" has logged out')
+        ;
+
+        $user = $this->mockClassWithProperties(BackendUser::class);
+        $user->username = 'foobar';
+
+        $token = $this->createMock(TokenInterface::class);
+        $token
+            ->expects($this->once())
+            ->method('getUser')
+            ->willReturn($user)
+        ;
+
+        $security = $this->createMock(Security::class);
+        $security
+            ->expects($this->once())
+            ->method('getToken')
+            ->willReturn($token)
+        ;
+
+        $response = new Response();
+
+        $event = new LogoutEvent(new Request(), null);
+        $event->setResponse($response);
+
+        $listener = $this->mockLogoutSuccessListener(null, null, $framework, $security, $logger);
+        $listener($event);
+    }
+
+    public function testDoesNotAddALogEntryIfTheUserIsNotSupported(): void
+    {
+        $framework = $this->mockContaoFramework();
+
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects($this->never())
+            ->method('info')
+        ;
+
+        $token = $this->createMock(TokenInterface::class);
+        $token
+            ->expects($this->once())
+            ->method('getUser')
+            ->willReturn($this->createMock(UserInterface::class))
+        ;
+
+        $security = $this->createMock(Security::class);
+        $security
+            ->expects($this->once())
+            ->method('getToken')
+            ->willReturn($token)
+        ;
+
+        $response = new Response();
+
+        $event = new LogoutEvent(new Request(), null);
+        $event->setResponse($response);
+
+        $listener = $this->mockLogoutSuccessListener(null, null, $framework, $security, $logger);
+        $listener($event);
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testTriggersThePostLogoutHook(): void
+    {
+        $this->expectDeprecation('Since contao/core-bundle 4.5: Using the "postLogout" hook has been deprecated %s.');
+
+        $user = $this->mockClassWithProperties(BackendUser::class);
+        $user->username = 'foobar';
+
+        $systemAdapter = $this->mockAdapter(['importStatic']);
+        $systemAdapter
+            ->expects($this->once())
+            ->method('importStatic')
+            ->with(static::class)
+            ->willReturn($this)
+        ;
+
+        $framework = $this->mockContaoFramework([System::class => $systemAdapter]);
+        $framework
+            ->expects($this->once())
+            ->method('initialize')
+        ;
+
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger
+            ->expects($this->once())
+            ->method('info')
+            ->with('User "foobar" has logged out')
+        ;
+
+        $token = $this->createMock(TokenInterface::class);
+        $token
+            ->expects($this->once())
+            ->method('getUser')
+            ->willReturn($user)
+        ;
+
+        $security = $this->createMock(Security::class);
+        $security
+            ->expects($this->once())
+            ->method('getToken')
+            ->willReturn($token)
+        ;
+
+        $response = new Response();
+
+        $event = new LogoutEvent(new Request(), null);
+        $event->setResponse($response);
+
+        $GLOBALS['TL_HOOKS']['postLogout'][] = [static::class, 'onPostLogout'];
+
+        $listener = $this->mockLogoutSuccessListener(null, null, $framework, $security, $logger);
+        $listener($event);
+
+        unset($GLOBALS['TL_HOOKS']);
+    }
+
+    public function onPostLogout(): void
+    {
+        // Dummy method to test the postLogout hook
+    }
+
+    public function testRemovesTargetPathFromSessionWithUsernamePasswordToken(): void
+    {
+        $session = $this->createMock(SessionInterface::class);
+        $session
+            ->expects($this->once())
+            ->method('remove')
+            ->with('_security.contao_frontend.target_path')
+        ;
+
+        $request = new Request();
+        $request->request->set('_target_path', 'http://localhost/home');
+        $request->setSession($session);
+
+        $httpUtils = $this->createMock(HttpUtils::class);
+        $httpUtils
+            ->expects($this->once())
+            ->method('createRedirectResponse')
+            ->with($request, 'http://localhost/home')
+            ->willReturn(new RedirectResponse('http://localhost/home'))
+        ;
+
+        $scopeMatcher = $this->createMock(ScopeMatcher::class);
+        $scopeMatcher
+            ->expects($this->once())
+            ->method('isBackendRequest')
+            ->willReturn(false)
+        ;
+
+        $token = $this->createMock(UsernamePasswordToken::class);
+        $token
+            ->expects($this->once())
+            ->method('getFirewallName')
+            ->willReturn('contao_frontend')
+        ;
+
+        $security = $this->createMock(Security::class);
+        $security
+            ->expects($this->once())
+            ->method('getToken')
+            ->willReturn($token)
+        ;
+
+        $response = new Response();
+
+        $event = new LogoutEvent($request, null);
+        $event->setResponse($response);
+
+        $listener = $this->mockLogoutSuccessListener($httpUtils, $scopeMatcher, $this->mockContaoFramework(), $security);
+        $listener($event);
+    }
+
+    public function testRemovesTargetPathFromSessionWithTwoFactorToken(): void
+    {
+        $session = $this->createMock(SessionInterface::class);
+        $session
+            ->expects($this->once())
+            ->method('remove')
+            ->with('_security.contao_frontend.target_path')
+        ;
+
+        $request = new Request();
+        $request->request->set('_target_path', 'http://localhost/home');
+        $request->setSession($session);
+
+        $httpUtils = $this->createMock(HttpUtils::class);
+        $httpUtils
+            ->expects($this->once())
+            ->method('createRedirectResponse')
+            ->with($request, 'http://localhost/home')
+            ->willReturn(new RedirectResponse('http://localhost/home'))
+        ;
+
+        $scopeMatcher = $this->createMock(ScopeMatcher::class);
+        $scopeMatcher
+            ->expects($this->once())
+            ->method('isBackendRequest')
+            ->willReturn(false)
+        ;
+
+        $token = $this->createMock(TwoFactorToken::class);
+        $token
+            ->expects($this->once())
+            ->method('getFirewallName')
+            ->willReturn('contao_frontend')
+        ;
+
+        $security = $this->createMock(Security::class);
+        $security
+            ->expects($this->once())
+            ->method('getToken')
+            ->willReturn($token)
+        ;
+
+        $response = new Response();
+
+        $event = new LogoutEvent($request, null);
+        $event->setResponse($response);
+
+        $listener = $this->mockLogoutSuccessListener($httpUtils, $scopeMatcher, $this->mockContaoFramework(), $security);
+        $listener($event);
+    }
+
+    private function mockLogoutSuccessListener(HttpUtils $httpUtils = null, ScopeMatcher $scopeMatcher = null, ContaoFramework $framework = null, Security $security = null, LoggerInterface $logger = null): LogoutSuccessListener
+    {
+        if (null === $httpUtils) {
+            $httpUtils = $this->createMock(HttpUtils::class);
+        }
+
+        if (null === $scopeMatcher) {
+            $scopeMatcher = $this->createMock(ScopeMatcher::class);
+        }
+
+        if (null === $framework) {
+            $framework = $this->mockContaoFramework();
+        }
+
+        if (null === $security) {
+            $security = $this->createMock(Security::class);
+            $security
+                ->expects($this->once())
+                ->method('getToken')
+            ;
+        }
+
+        return new LogoutSuccessListener($httpUtils, $scopeMatcher, $framework, $security, $logger);
     }
 }
